requirements.txt
pyyaml
mechanize
beautifulsoup4